cast to float for macOs

+++ ntt/test/ntt_test.h
@@ -139,7 +139,7 @@ void init_tensor(TTensor &tensor, T start = static_cast<T>(0),
             // std::cout << ") = " << tensor(index) << std::endl;
         });
     } else if constexpr (std::is_same_v<T, _Float16>) {
-        std::uniform_real_distribution<_Float16> dis(start, stop);
+        std::uniform_real_distribution<float> dis(start, stop);
         ntt::apply(tensor.shape(), [&](auto &index) {
             tensor(index) = static_cast<_Float16>(dis(gen));
         });
